# docs/api/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
...
  /api/notifications/more-notifications:
    get:
      tags:
      - rest-notification-controller
      summary: Get some of the notification of the current user
      operationId: getSomeNotifications
      parameters:
      - name: from
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Notifications obtained
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification_Basic'
        "202":
          description: Notifications not found
  /api/notifications/more-mentions:
    get:
      tags:
      - rest-notification-controller
      summary: Get some of the mentions of the current user
      operationId: getSomeMentions
      parameters:
      - name: from
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "202":
          description: Mentions not found
        "200":
          description: Mentions obtained
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification_Basic'
...
  /api/notifications/delete-notification:
    delete:
      tags:
      - rest-notification-controller
      summary: Deletes a notification to a user
      operationId: deleteNotification
      parameters:
      - name: idTweet
        in: query
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: string
      responses:
        "200":
          description: Notification Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification_Basic'
        "400":
          description: Cannot delete the notification
components:
  schemas:
...
    Tweet_Basic:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/User_Basic'
        publishDate:
          type: string
          format: date-time
        likes:
          type: array
          items:
            $ref: '#/components/schemas/User_Basic'
        retweets:
          type: array
          items:
            $ref: '#/components/schemas/User_Basic'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Tweet_Basic'
        text:
          type: string
        media1:
          type: object
          properties:
            binaryStream:
              type: object
        media2:
          type: object
          properties:
            binaryStream:
              type: object
        media3:
          type: object
          properties:
            binaryStream:
              type: object
        media4:
          type: object
          properties:
            binaryStream:
              type: object
    User_Basic:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        nickname:
          type: string
        password:
          type: string
        profilePicture:
          type: object
          properties:
            binaryStream:
              type: object
        role:
          type: string
          enum:
          - ANONYMOUS
          - USER
          - ADMIN
        followers:
          type: array
          items:
            $ref: '#/components/schemas/User_Basic'
        followed:
          type: array
          items:
            $ref: '#/components/schemas/User_Basic'
        bookmarks:
          type: array
          items:
            $ref: '#/components/schemas/Tweet_Basic'
        joinDate:
          type: string
          format: date
        type:
          type: string
        verificationCode:
          type: string
        resetPasswordToken:
          type: string
        loggedIn:
          type: boolean
        enabled:
          type: boolean
        followedNumber:
          type: integer
          format: int32
        followersNumber:
          type: integer
          format: int32
        images:
          type: array
          writeOnly: true
          items:
            type: string
...
    Notification_Basic:
      type: object
      properties:
        id:
          type: integer
          format: int64
        tweetTrigger:
          $ref: '#/components/schemas/Tweet_Basic'
        userToNotify:
          $ref: '#/components/schemas/User_Basic'
        userWhoNotifies:
          $ref: '#/components/schemas/User_Basic'
        date:
          type: string
          format: date-time
        type:
          type: string
...